@@ -1,70 +1,68 @@
 {
-  "name": "koa-swagger-decorator",
-  "version": "1.8.5",
-  "description": "using decorator to automatically generate swagger doc for koa-router",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "dependencies": {
-    "@koa/router": "^10.1.1",
-    "@types/koa__router": "^8.0.11",
-    "@types/ramda": "^0.25.51",
-    "@types/validator": "^13.7.1",
-    "globby": "^11.0.4",
-    "is-type-of": "^1.2.1",
-    "koa-compose": "^4.1.0",
-    "koa2-swagger-ui": "^5.3.0",
-    "ramda": "^0.27.1",
-    "validator": "^13.7.0"
-  },
-  "devDependencies": {
-    "@commitlint/cli": "^8.3.6",
-    "@commitlint/config-conventional": "^8.3.6",
-    "@koa/cors": "^3.1.0",
-    "@koa/multer": "^3.0.0",
-    "@types/jest": "26",
-    "@types/koa": "^2.13.4",
-    "@types/koa-bodyparser": "^5.0.2",
-    "@types/koa__cors": "^3.1.1",
-    "@types/koa__multer": "^2.0.4",
-    "@types/mocha": "^5.2.7",
-    "@types/node": "^16.11.17",
-    "chai": "^4.3.4",
-    "fast-glob": "^3.2.7",
-    "husky": "^4.3.8",
-    "jest": "26",
-    "koa": "^2.13.4",
-    "koa-bodyparser": "^4.3.0",
-    "koa-static": "^4.0.3",
-    "multer": "^1.4.4",
-    "nodemon": "^2.0.15",
-    "power-assert": "^1.6.1",
-    "standard-version": "^9.3.2",
-    "supertest": "^2.0.1",
-    "ts-jest": "26",
-    "ts-node": "^10.4.0",
-    "typescript": "^4.5.4"
-  },
-  "scripts": {
-    "build": "rm -rf dist && tsc",
-    "start": "ts-node example/main",
-    "test": "jest",
-    "cov": "jest --coverage",
-    "prepublish": "npm run build",
-    "release": "standard-version",
-    "latest": "npm publish --registry=https://registry.npmjs.org"
-  },
-  "husky": {
-    "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
-    }
-  },
-  "keywords": [
-    "decorator",
-    "swagger",
-    "koa",
-    "koa-router"
-  ],
-  "author": "cody",
-  "license": "MIT",
-  "repository": "https://github.com/Cody2333/koa-swagger-decorator"
+    "name": "koa-swagger-decorator2",
+    "version": "2.0.0",
+    "description": "using decorator to automatically generate swagger doc for koa-router",
+    "main": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "dependencies": {
+        "@koa/router": "^10.1.1",
+        "globby": "^11.1.0",
+        "is-type-of": "^1.2.1",
+        "koa-compose": "^4.1.0",
+        "koa2-swagger-ui": "^5.5.0",
+        "ramda": "^0.28.0",
+        "validator": "^13.7.0"
+    },
+    "devDependencies": {
+        "@commitlint/cli": "^16.2.3",
+        "@commitlint/config-conventional": "^16.2.1",
+        "@koa/cors": "^3.3.0",
+        "@koa/multer": "^3.0.0",
+        "@types/jest": "^27.4.1",
+        "@types/koa": "^2.13.4",
+        "@types/koa-bodyparser": "^5.0.2",
+        "@types/koa__cors": "^3.3.0",
+        "@types/koa__router": "^8.0.11",
+        "@types/node": "^17.0.25",
+        "chai": "^4.3.6",
+        "fast-glob": "^3.2.11",
+        "husky": "^7.0.4",
+        "jest": "27",
+        "koa": "^2.13.4",
+        "koa-body": "^5.0.0",
+        "koa-bodyparser": "^4.3.0",
+        "koa-static": "^5.0.0",
+        "multer": "^1.4.4",
+        "nodemon": "^2.0.15",
+        "prettier": "^2.6.2",
+        "standard-version": "^9.3.2",
+        "supertest": "6.2.2",
+        "ts-jest": "27",
+        "ts-node": "^10.7.0",
+        "typescript": "^4.6.3"
+    },
+    "scripts": {
+        "build": "rm -rf dist && tsc",
+        "start": "nodemon example/main.ts",
+        "lint": "prettier --write '**/*.{ts,json}'",
+        "test": "jest",
+        "cov": "jest --coverage",
+        "prepublish": "npm run build",
+        "release": "standard-version",
+        "latest": "npm publish --registry=https://registry.npmjs.org"
+    },
+    "husky": {
+        "hooks": {
+            "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
+        }
+    },
+    "keywords": [
+        "decorator",
+        "swagger",
+        "koa",
+        "koa-router"
+    ],
+    "author": "tomoat",
+    "license": "MIT",
+    "repository": "https://github.com/nestfiy/koa-swagger-decorator2"
 }
