@@ -1,6 +1,6 @@
 const common = {
-  baseUrl: 'http://localhost:3000',
-  port: 3000
-};
+    baseUrl: 'http://localhost:3000',
+    port: 3000
+}
 
-export default common;
+export default common
