@@ -1,35 +1,33 @@
-import Doc, { description } from '../../../lib'; // 2 import style avaliable
+import Doc, { description } from '../../../lib' // 2 import style avaliable
 
-const {
-  request, summary, query, tags, responses, prefix, header
-} = Doc;
+const { request, summary, query, tags, responses, prefix, header } = Doc
 
-const tag = tags(['prefix']);
+const tag = tags(['prefix'])
 
 @prefix('/v1')
 export default class PrefixRouter {
-  @request('GET', '/prefix')
-  @summary('class decorator for url prefix')
-  @description('class decorator for url prefix')
-  @tag
-  @header({
-    appKey: {
-      type: 'string',
-      description: 'app key'
+    @request('GET', '/prefix')
+    @summary('class decorator for url prefix')
+    @description('class decorator for url prefix')
+    @tag
+    @header({
+        appKey: {
+            type: 'string',
+            description: 'app key'
+        }
+    })
+    @query({
+        page: {
+            type: 'string',
+            description: 'page number'
+        }
+    })
+    @responses({
+        200: { description: 'success' },
+        500: { description: 'something wrong about server' }
+    })
+    static async classPrefix(ctx) {
+        const { page } = ctx.request.query
+        ctx.body = { result: page }
     }
-  })
-  @query({
-    page: {
-      type: 'string',
-      description: 'page number'
-    }
-  })
-  @responses({
-    200: { description: 'success' },
-    500: { description: 'something wrong about server' }
-  })
-  static async classPrefix(ctx) {
-    const { page } = ctx.request.query;
-    ctx.body = { result: page };
-  }
 }
