@@ -1,15 +1,14 @@
-import path from 'path';
-import { SwaggerRouter } from '../../lib';
+import path from 'path'
+import { SwaggerRouter } from '../../lib'
 
 // init router
-const router = new SwaggerRouter();
+const router = new SwaggerRouter()
 
 // load controllers
-router.mapDir(path.resolve(__dirname, '../routes'));
+router.mapDir(path.resolve(__dirname, '../routes'))
 
 // dump swagger json
 router.dumpSwaggerJson({
-  filename: 'swagger.json', // default is swagger.json
-  dir: process.cwd(), // default is process.cwd()
-});
-
+    filename: 'swagger.json', // default is swagger.json
+    dir: process.cwd() // default is process.cwd()
+})
