@@ -1,55 +1,61 @@
 /**
  * All other regular column types.
  */
-export type PropertyType = "string" | "number" | "boolean" | "array" | "object";
+export type PropertyType = 'string' | 'number' | 'boolean' | 'array' | 'object'
 
-type CustromStringFormatT = string;
-export type StringFormatT = "date" | "date-time" | "password" | "byte" | "binary" | CustromStringFormatT;
+type CustromStringFormatT = string
+export type StringFormatT =
+    | 'date'
+    | 'date-time'
+    | 'password'
+    | 'byte'
+    | 'binary'
+    | CustromStringFormatT
 
 class BasePropertyOptions {
-    required?: boolean;
-    example?: any;
-    description?: string;
-    readOnly?: boolean;
-    writeOnly?: boolean;
-    nullable?: boolean;
+    required?: boolean
+    example?: any
+    description?: string
+    readOnly?: boolean
+    writeOnly?: boolean
+    nullable?: boolean
 }
 
 type StringPropertyOptionsT = BasePropertyOptions & {
-    type: "string";
-    minLength?: number;
-    maxLength?: number;
-    format?: StringFormatT;
-    pattern?: string;
-};
+    type: 'string'
+    minLength?: number
+    maxLength?: number
+    format?: StringFormatT
+    pattern?: string
+}
 
 type NumberPropertyOptionsT = BasePropertyOptions & {
-    type: "number";
-    minimum?: number;
-    exclusiveMinimum?: boolean;
-    maximum?: number;
-    exclusiveMaximum?: boolean;
-    multipleOf?: number;
-};
+    type: 'number'
+    minimum?: number
+    exclusiveMinimum?: boolean
+    maximum?: number
+    exclusiveMaximum?: boolean
+    multipleOf?: number
+}
 
 type BooleanPropertyOptionsT = BasePropertyOptions & {
-    type: "boolean";
-};
+    type: 'boolean'
+}
 
 type ArrayPropertyOptionsT = BasePropertyOptions & {
-    type: "array",
-    minItems?: number;
-    maxItems?: number;
-    uniqueItems?: boolean;
-    items?: PropertyOptions;
-};
+    type: 'array'
+    minItems?: number
+    maxItems?: number
+    uniqueItems?: boolean
+    items?: PropertyOptions
+}
 
 type ObjectPropertyOptionsT = BasePropertyOptions & {
-    type: "object",
-    properties?: any;
-    minProperties?: number;
-    maxProperties?: number;
-};
+    type: 'object'
+    properties?: any
+    minProperties?: number
+    maxProperties?: number
+}
 
 export type PropertyOptions =
     | StringPropertyOptionsT
@@ -57,7 +63,6 @@ export type PropertyOptions =
     | BooleanPropertyOptionsT
     | ArrayPropertyOptionsT
     | ObjectPropertyOptionsT
-;
 
 /**
  *
@@ -65,47 +70,59 @@ export type PropertyOptions =
  */
 function deepClone(source: any) {
     if (!source || typeof source !== 'object') {
-        return null;
+        return null
     }
-    const targetObj: any = source.constructor === Array ? [] : {};
+    const targetObj: any = source.constructor === Array ? [] : {}
     for (const keys in source) {
         if (source.hasOwnProperty(keys)) {
             if (source[keys] && typeof source[keys] === 'object') {
-                targetObj[keys] = source[keys].constructor === Array ? [] : {};
-                targetObj[keys] = deepClone(source[keys]);
+                targetObj[keys] = source[keys].constructor === Array ? [] : {}
+                targetObj[keys] = deepClone(source[keys])
             } else {
-                targetObj[keys] = source[keys];
+                targetObj[keys] = source[keys]
             }
         }
     }
-    return targetObj;
+    return targetObj
 }
 
 /**
  * Made for empty class
  */
 export function swaggerClass(): Function {
-    return function (target: any, propertyKey: string, descriptor: PropertyDescriptor) {
-        if (target.swaggerDocument == undefined) target.swaggerDocument = {};
-        if (target.swaggerClass == undefined) target.swaggerClass = target;
+    return function (
+        target: any,
+        propertyKey: string,
+        descriptor: PropertyDescriptor
+    ) {
+        if (target.swaggerDocument == undefined) target.swaggerDocument = {}
+        if (target.swaggerClass == undefined) target.swaggerClass = target
         if (target.swaggerClass != target) {
-            target.swaggerClass = target;
-            target.swaggerDocument = deepClone(target.swaggerDocument);
+            target.swaggerClass = target
+            target.swaggerDocument = deepClone(target.swaggerDocument)
         }
-    };
+    }
 }
 
 /**
  * @param options
  */
 export function swaggerProperty(options?: PropertyOptions): Function {
-    return function (target: any, propertyKey: string, descriptor: PropertyDescriptor) {
-        if (target.constructor.swaggerDocument == undefined) target.constructor.swaggerDocument = {};
-        if (target.constructor.swaggerClass == undefined) target.constructor.swaggerClass = target.constructor;
+    return function (
+        target: any,
+        propertyKey: string,
+        descriptor: PropertyDescriptor
+    ) {
+        if (target.constructor.swaggerDocument == undefined)
+            target.constructor.swaggerDocument = {}
+        if (target.constructor.swaggerClass == undefined)
+            target.constructor.swaggerClass = target.constructor
         if (target.constructor.swaggerClass != target.constructor) {
-            target.constructor.swaggerClass = target.constructor;
-            target.constructor.swaggerDocument = deepClone(target.constructor.swaggerDocument);
+            target.constructor.swaggerClass = target.constructor
+            target.constructor.swaggerDocument = deepClone(
+                target.constructor.swaggerDocument
+            )
         }
-        target.constructor.swaggerDocument[propertyKey] = options;
-    };
+        target.constructor.swaggerDocument[propertyKey] = options
+    }
 }
